rabbitmq/requirements.txt
backoff==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
pamqp==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4.0"
aiormq==6.7.6 ; python_version >= "3.10" and python_version < "4.0"
aio-pika==9.1.2 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.9.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.8 ; python_version >= "3.10" and python_version < "4.0"